Remove buggy compose marker clustering
