+++ app/src/main/java/com/github/geohunt/app/maps/GoogleMapDisplay.kt
@@ -6,8 +6,8 @@ import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
+import com.github.geohunt.app.maps.marker.DisplayMarkers
 import com.github.geohunt.app.maps.marker.Marker
-import com.github.geohunt.app.maps.marker.MarkerDisplay
 import com.github.geohunt.app.model.database.api.Location
 import com.github.geohunt.app.ui.screens.home.HomeViewModel
 import com.github.geohunt.app.ui.screens.maps.MapsViewModel
@@ -23,7 +23,6 @@ import kotlin.math.pow
 
 private val mockBitmap: Bitmap = Bitmap.createBitmap(100, 100, Bitmap.Config.ARGB_8888)
 private val epflCoordinates = LatLng(46.51958, 6.56398)
-private var challengeDatabase: List<Marker> = mutableListOf()
 
 /*
  * Displays the Google Map and its content
@@ -59,8 +58,6 @@ fun GoogleMapDisplay(
             val longitude = coordinateCenter.longitude
 
             val radius = 38000 / 2.0.pow((zoom - 3).toDouble()) * cos(latitude * PI / 180)
-            Log.d("MAPS SEEN CHALLENGES RAD", radius.toString())
-            //val radius = 30.0 //TODO change this value later
             val location = Location(latitude, longitude)
             val neighboringSectors = location.getNeighboringSectors(radius)
 
@@ -75,55 +72,20 @@ fun GoogleMapDisplay(
                 Log.d("MAPS SEEN CHALLENGES", it.toString())
 
                 val marker = Marker(
-                    markerPosition = LatLng(it.location?.latitude ?: 0.0, it.location?.longitude ?: 0.0),
-                    markerTitle = it.id ?: "",
-                    markerSnippet = it.description ?: "",
+                    title = it.id ?: "",
                     //TODO use string url a bit later
                     image = mockBitmap,
+                    coordinates = LatLng(it.location?.latitude ?: 0.0, it.location?.longitude ?: 0.0),
                     expiryDate = it.expirationDate ?: LocalDateTime.of(2024, Month.MAY, 1, 19, 0),
                 )
 
                 markersList.add(marker)
             }
 
-            MarkerDisplay(items = markersList)
+            DisplayMarkers(markers = markersList)
 
             content()
 
         }
     }
-}
-
-/*
- * Loads the challenges from the database
- * If the database is empty, it creates a list of mock markers
- *
- * @param markers a list of markers
- */
-fun loadChallenges(markers: List<Marker>){
-    challengeDatabase = markers.ifEmpty {
-        createListOfMockMarkers()
-    }
-}
-
-/*
- * Creates a list of mock markers that are used
- * to populate the map with markers
- *
- * @return a list of mock markers
- */
-private fun createListOfMockMarkers(): List<Marker> {
-    val mockChallengeDatabase =  mutableListOf<Marker>()
-
-    for (i in 1..100) {
-        mockChallengeDatabase.add(Marker(
-            LatLng(46.51958 + i * 0.01, 6.56398 + i * 0.01),
-            "Event $i",
-            "Expires on 1 May 2024 at 19:39",
-            mockBitmap,
-            LocalDateTime.of(2024, Month.MAY, 1, 19, 39, 12))
-        )
-    }
-
-    return mockChallengeDatabase
 }
