--- app/src/main/java/com/github/geohunt/app/maps/marker/MarkerData.kt
@@ -1,11 +1,0 @@
-package com.github.geohunt.app.maps.marker
-
-import android.graphics.Bitmap
-import com.google.android.gms.maps.model.LatLng
-import java.time.LocalDateTime
-
-data class MarkerData(
-    val title: String,
-    val image: Bitmap,
-    val coordinates: LatLng,
-    val expiryDate: LocalDateTime)
