+++ app/src/main/java/com/github/geohunt/app/maps/marker/MarkerDisplay.kt
@@ -2,9 +2,12 @@ package com.github.geohunt.app.maps.marker
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.*
@@ -12,6 +15,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
@@ -22,117 +26,68 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.github.geohunt.app.R
 import com.github.geohunt.app.ui.theme.geoHuntRed
+import com.google.android.gms.maps.model.BitmapDescriptorFactory
 import com.google.maps.android.compose.*
 import com.google.maps.android.compose.clustering.Clustering
 
-/*
- * Displays the markers on the map
- * and clusters them when they are close to each other
+/**
+ * Displays provided markers on the map
  *
- * @param items The list of markers to display
+ * @ param markers The list of markers to display
  */
-@OptIn(MapsComposeExperimentalApi::class)
 @Composable
-fun MarkerDisplay(items: List<Marker>) {
-    var selectedMarker: Marker? by remember { mutableStateOf(null) }
-    val isMarkerClicked = remember { mutableStateOf(false) }
+fun DisplayMarkers(markers: List<Marker>) {
 
-    DisplayMarkers(markers = items)
-
-
-    /*Clustering(
-        items = items,
-
-        // When clicked, update the selected marker
-        // and set the markerClicked state to true
-        onClusterItemClick = { marker ->
-            selectedMarker = marker
-            isMarkerClicked.value = true
-
-            true
-        },
-
-        // Render the non clustered markers
-        // and info window if the marker is clicked
-        clusterItemContent = {marker ->
-            if (isMarkerClicked.value) {
-                if (selectedMarker == marker) {
-                    MarkerInfoBox(marker = marker)
-                }
-            }
-            Image(
-                painter = painterResource(id = R.drawable.marker),
-                contentDescription = "Marker Icon for ${marker.title}",
+    markers.forEach { challenge ->
+        MarkerInfoWindowContent(
+            state = rememberMarkerState(position = challenge.coordinates),
+            title = challenge.title,
+            snippet = challenge.expiryDate.toString(),
+            icon = BitmapDescriptorFactory.defaultMarker(geoHuntRed.red)
+        ) {
+            Box(
                 modifier = Modifier
-                    .size(50.dp)
-                    .alpha(if (selectedMarker == marker) 0f else 1f)
-                    // Clickable modifier to make the marker clickable for the tests
-                    .clickable(onClick = {})
-            )
-        },
-
-        // Render the clustered markers
-        clusterContent = { cluster ->
-            val clusterSize = cluster.size
-
-            Surface(
-                modifier = Modifier.size(determineClusterSize(clusterSize)).testTag("Cluster for $clusterSize markers"),
-                shape = CircleShape,
-                color = determineClusterColor(clusterSize),
-                contentColor = Color.Black,
-                border = BorderStroke(4.dp, Color.Green)
+                    .background(
+                        color = MaterialTheme.colorScheme.onPrimary,
+                        shape = RoundedCornerShape(35.dp, 35.dp, 35.dp, 35.dp)
+                    )
             ) {
-                Box(contentAlignment = Alignment.Center) {
+                Column(
+                    horizontalAlignment = Alignment.CenterHorizontally
+                ) {
+                    //The image displayed at the top of the info window
+                    Image(
+                        painter = painterResource(id = R.drawable.radar_icon),
+                        contentDescription = "Radar Icon",
+                        modifier = Modifier
+                            .size(90.dp)
+                            .padding(top = 16.dp),
+                        contentScale = ContentScale.Crop)
+
+                    Spacer(modifier = Modifier.height(16.dp))
+
+                    //The middle text containing the title of the challenge
                     Text(
-                        "%,d".format(clusterSize),
-                        fontSize = determineFontSize(clusterSize),
-                        fontWeight = FontWeight.Black,
+                        text = challenge.title,
                         textAlign = TextAlign.Center,
                         modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(horizontal = 8.dp)
-                        )
+                            .fillMaxWidth(),
+                        style = MaterialTheme.typography.displayMedium,
+                    )
+
+                    //The bottom text containing the expiry date of the challenge
+                    Text(
+                        text = challenge.expiryDate.toString(),
+                        textAlign = TextAlign.Center,
+                        modifier = Modifier
+                            .padding(top = 10.dp, start = 25.dp, end = 25.dp)
+                            .fillMaxWidth(),
+                        style = MaterialTheme.typography.headlineSmall,
+                    )
+
+                    Spacer(modifier = Modifier.height(16.dp))
                 }
             }
         }
-    )*/
-}
-
-/*
- * Determines the color of the cluster based on the size of the cluster
- *
- * @param clusterSize The size of the cluster
- */
-private fun determineClusterColor(clusterSize: Int): Color {
-    return when {
-        clusterSize < 10 -> Color.Green
-        clusterSize < 100 -> Color.Yellow
-        else -> geoHuntRed
-    }
-}
-
-/*
- * Determines the diameter of the cluster based on the size of the cluster
- *
- * @param clusterSize The size of the cluster
- */
-private fun determineClusterSize(clusterSize: Int): Dp {
-    return when {
-        clusterSize < 10 -> 40.dp
-        clusterSize < 100 -> 60.dp
-        else -> 80.dp
-    }
-}
-
-/*
- * Determines the font size of the cluster based on the size of the cluster
- *
- * @param clusterSize The size of the cluster
- */
-private fun determineFontSize(clusterSize: Int): TextUnit {
-    return when {
-        clusterSize < 10 -> 16.sp
-        clusterSize < 100 -> 20.sp
-        else -> 26.sp
     }
 }
